+++ src/db/dbcard.c
@@ -80,7 +80,7 @@ int GetCardPriceModifier(unsigned int card_id)
 
 }
 
-float GetCardModifierLength(unsigned int card_id)
+unsigned int GetCardModifierLength(unsigned int card_id)
 {
 
     Card *temp = (Card *)cards->elements;
